Moved methods from Actions to ToolUtils
