+++ Sandbox/.agentforge/settings/models.yaml
@@ -2,14 +2,14 @@
 ModelSettings:
 #  API: gemini_api
 #  Model: gemini-flash
-  API: openai_api
-  Model: omni_model
+#  API: openai_api
+#  Model: omni_model
 #  API: claude3_api
 #  Model: claude-3
 #  API: ollama_api
 #  Model: DolphinMistral
-#  API: lm_studio_api
-#  Model: LMStudio
+  API: lm_studio_api
+  Model: LMStudio
   Params: # Default parameter values
     max_new_tokens: 3000
     temperature: 0.8
