+++ src/agentforge/tools/GetText.py
@@ -24,7 +24,8 @@ class GetText:
                 elif file_name_or_url.endswith('.txt') or file_name_or_url.endswith('.md'):
                     return self.read_txt(file_name_or_url)
                 else:
-                    return {"status": "failure", "content": "Unsupported file format - Use URL, PDF or TXT formats."}
+                    return {"status": "failure", "content": "Unsupported file format - Use URL, PDF, TXT or Markdown "
+                                                            "formats."}
         except Exception as e:
             return {"status": "failure", "content": str(e)}
 
@@ -45,7 +46,7 @@ class GetText:
             with open(filename, 'rb') as file:
                 content = io.BytesIO(file.read())
                 text = self.extract_text_from_pdf(content)
-            return {"status": "success", "content": text['content']}
+            return text
         except Exception as e:
             return {"status": "failure", "content": str(e)}
 
@@ -117,49 +118,4 @@ if __name__ == "__main__":
     gettext_instance = GetText()
     filename_or_url = 'Documents'  # Replace with your file path or URL
     file_content = gettext_instance.read_file(filename_or_url)
-    print(file_content)
-
-
-# import pypdf
-# import requests
-# import io
-#
-#
-# class GetText:
-#     def read_file(self, file_name_or_url):
-#         if file_name_or_url.startswith('http://') or file_name_or_url.startswith('https://'):
-#             return self.read_from_url(file_name_or_url)
-#         else:
-#             if file_name_or_url.endswith('.pdf'):
-#                 return self.read_pdf(file_name_or_url)
-#             elif file_name_or_url.endswith('.txt'):
-#                 return self.read_txt(file_name_or_url)
-#             else:
-#                 return "Unsupported file format"
-#
-#     def read_pdf(self, filename):
-#         with open(filename, 'rb') as file:
-#             text = self.extract_text_from_pdf(file)
-#         return text
-#
-#     @staticmethod
-#     def read_txt(filename):
-#         with open(filename, 'r', encoding='utf-8') as file:
-#             return file.read()
-#
-#     def read_from_url(self, url):
-#         response = requests.get(url)
-#         if url.endswith('.pdf'):
-#             return self.extract_text_from_pdf(io.BytesIO(response.content))
-#         elif url.endswith('.txt'):
-#             return response.text
-#         else:
-#             return "Unsupported file format"
-#
-#     @staticmethod
-#     def extract_text_from_pdf(file_stream):
-#         text = ""
-#         reader = pypdf.PdfReader(file_stream)
-#         for page in reader.pages:
-#             text += page.extract_text()
-#         return text
+    print(file_content['content'])
